+++ commands/__init__.py
@@ -3,14 +3,14 @@ import sublime
 
 st_version = int(sublime.version())
 if st_version > 3000:
-    from JoomlaPack.commands.project import NewProjectCommand
+    # from JoomlaPack.commands.project import NewProjectCommand
     from JoomlaPack.commands.component import NewComponentCommand
     from JoomlaPack.commands.package import NewPackageCommand
     from JoomlaPack.commands.plugin import NewPluginCommand
     from JoomlaPack.commands.plugin import AddFormToPluginCommand
     from JoomlaPack.commands.plugin import AddFieldToPluginCommand
 else:
-    from commands.project import NewProjectCommand
+    # from commands.project import NewProjectCommand
     from commands.component import NewComponentCommand
     from commands.package import NewPackageCommand
     from commands.plugin import NewPluginCommand
@@ -18,7 +18,7 @@ else:
     from commands.plugin import AddFieldToPluginCommand
 
 __all__ = [
-    'NewProjectCommand',
+    # 'NewProjectCommand',
 
     'NewComponentCommand',