+++ lib/__init__.py
@@ -6,19 +6,22 @@ if st_version > 3000:
     from JoomlaPack.lib.file import File
     from JoomlaPack.lib.folder import Folder
     from JoomlaPack.lib.helper import Helper
-    from JoomlaPack.lib.json_file import JsonFile
+    from JoomlaPack.lib.json import Json
     from JoomlaPack.lib.project import Project
+    from JoomlaPack.lib.xml import Xml
 else:
     from lib.file import File
     from lib.folder import Folder
     from lib.helper import Helper
-    from lib.json_file import JsonFile
+    from lib.json import Json
     from lib.project import Project
+    from lib.xml import Xml
 
 __all__ = [
     'File',
     'Folder',
     'Helper',
-    'JsonFile',
-    'Project'
+    'Json',
+    'Project',
+    'Xml'
 ]